requirements.txt
alembic==1.3.1
Click==7.0
Flask==1.1.1
Flask-Admin==1.5.4
Flask-FontAwesome==0.1.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
WTForms==2.2.1